[Silver III] Title: 주유소, Time: 88 ms, Memory: 3588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -7,8 +7,8 @@ int main()
 {
 	int lineNum;
 	cin >> lineNum;
-	vector<int> lv(lineNum - 1);
-	vector<int> sv(lineNum);
+	vector<long> lv(lineNum - 1);
+	vector<long> sv(lineNum);
 
 	for (int i = 0; i < lineNum - 1; i++)
 	{
@@ -19,8 +19,8 @@ int main()
 		cin >> sv[i];
 	}
 
-	int minCost = 1000000001;
-	int result = 0;
+	long minCost = 1000000001;
+	long long result = 0;
 
 	for (int i = 0; i < lineNum - 1; i++) 
 	{
